[OWL-1695] Make tests well organized.
